./autogen.sh runs auto* and then ./configure
It should be all you need to do to get working Makefiles on your
platform, whatever your platform is. :)

Let me know if it doesn't generate Makefiles for you. There will still
be some bugs to work out in detecting openssl correctly, in checking for
only the right header files, etc. But we're on our way.


git-svn-id: https://svn.torproject.org/svn/tor/trunk@15 55e972cd-5a19-0410-ae62-a4d7a52db4cd