clean up logging, allow user to specify log files
If DebugLogFile is specified, log to it at -l debug
If LogFile is specified, log to it at the -l from the commandline
  (default info)
If no LogFile *and* not a Daemon, then log to stdout.
Make conn->s = -1 by default (this might break things)
When kill -USR1, prefer to log at INFO, but make sure they always see it.


git-svn-id: https://svn.torproject.org/svn/tor/trunk@596 55e972cd-5a19-0410-ae62-a4d7a52db4cd