send the end cell when we realize we're going to end, not when we're closing the stream.
this lets us put a payload in the end cell if we want to,
to describe why we're closing the stream.

there are still some places where we don't send the end cell
immediately. i need to track them down. but it's a low priority,
since i've made it send the end cell when we close the stream if
we haven't already sent it.


git-svn-id: https://svn.torproject.org/svn/tor/trunk@640 55e972cd-5a19-0410-ae62-a4d7a52db4cd