added User and Group options -- if you set them, tor will try to setuid and setgid respectively, and die if it can't.
(If the User option is set, tor will setgid to the user's gid as well.)

This happens after the pidfile is created, so that in cases where tor
needs to be root to work with the pidfile, it will at least be able to
create it, although it won't be able to delete it.  That sucks, but
it's somewhat better than not being able to create the pidfile in the
first place.


git-svn-id: https://svn.torproject.org/svn/tor/trunk@652 55e972cd-5a19-0410-ae62-a4d7a52db4cd