Fix assertion failure spotted by arma.
Apparently, when a DNS failure was already cached, then when we tried
to mark the exit connection as closed, we'd try to remove it from the
pending queue anyway, and hit an assert.  Now, we put failed-resolve
connections in a separate state so that mark_for_close does the right
thing.


git-svn-id: https://svn.torproject.org/svn/tor/trunk@1196 55e972cd-5a19-0410-ae62-a4d7a52db4cd