The crowning bugfix.
The problem was that with high load, circuit package window was
reaching 0. Whenever we got a circuit-level sendme, we were
reading a lot on each socket, but only writing out a bit. So we
would eventually reach eof. This would be noticed and acted on
even when there are still bytes sitting in the inbuf.


git-svn-id: https://svn.torproject.org/svn/tor/trunk@2932 55e972cd-5a19-0410-ae62-a4d7a52db4cd