Note a vulnerability with our current recommended-version concensus building.
Make the warnings about invalid and unnamed nodes scale better, and
update the text of the warnings.

Change router_have_minimum_dir_info() to only be happy when it has
enough network-statuses ("more than half") to be willing to actually
build circuits.

Not yet done: when we fail to get a networkstatus that we wanted, and
!router_have_minimum_dir_info(), we should retry it quicker than a whole
minute from now.


git-svn-id: https://svn.torproject.org/svn/tor/trunk@6227 55e972cd-5a19-0410-ae62-a4d7a52db4cd