make it work with base64
base64 from coreutils is really picky about the number of equation
marks (=) that end a base64 encoded string.  Unfortunately Tor doesn't
believe they are important at all so it cuts them all off all the time.

For perl purposes we just added a good measure of = and were happy about
it.  For base64 we now need to add the proper number of equation marks.
It turns out that number is 1 for the length of strings we process.

git-svn-id: https://svn.torproject.org/svn/tor/trunk@18170 55e972cd-5a19-0410-ae62-a4d7a52db4cd